Cleaning up!

+++ ann_1.1.2/include/ANN/ANN.h
@@ -155,7 +155,7 @@ enum ANNbool {ANNfalse = 0, ANNtrue = 1}; // ANN boolean type (non ANSI C++)
 //		not occur in distance calculation.
 //----------------------------------------------------------------------
 
-typedef double	ANNcoord;				// coordinate data type
+typedef float		ANNcoord;				// coordinate data type
 typedef double	ANNdist;				// distance data type
 
 //----------------------------------------------------------------------
